First draft of full render section in implementation chapter.

+++ gen-gif.sh
@@ -1,10 +1,17 @@
 #!/usr/bin/env sh
 
-cd images || exit
-rm -f tea.gif
-../build/bin/release/volumetric-ray-tracer -f ../test-objects/teapot.obj -t32 -q --tiles 32 --frames 24 -o img.png -w 512
-ffmpeg -f image2 -framerate 12 -i img_%d.png -vf scale=1024x1024 gif.avi
-ffmpeg -i gif.avi -pix_fmt rgb24 "$ARCH"_tea.gif
+BASEDIR=$(dirname "$0")
+
+if [ $# -lt 2 ]; then
+    echo "usage: gen-gif.sh <.obj> <.gif>"
+    exit 1
+fi
+
+mkdir -p tmp
+
+rm -f "$2"
+"$BASEDIR"/build/bin/release/volumetric-ray-tracer -f "$1" -t32 -q --tiles 32 --frames 24 -o ./tmp/img.png -w 512
+ffmpeg -f image2 -framerate 12 -i ./tmp/img_%d.png -vf scale=1024x1024 gif.avi
+ffmpeg -i gif.avi -pix_fmt rgb24 "$2"
 rm gif.avi
-rm -rf img*
-cd ..
+rm -rf tmp
